+++ react/ProductSummaryCustom.tsx
@@ -170,9 +170,7 @@ function ProductSummaryCustom({
         to: href,
         onClick: autocompleteSummary ? actionOnClick : undefined,
         onClickCapture: autocompleteSummary ? undefined : actionOnClick,
-        'aria-label': `View product details for ${
-          product?.productName || 'product'
-        }`,
+        'aria-label': `${product?.productName || 'product'}`,
       }
     : {
         page: 'store.product',
@@ -184,9 +182,7 @@ function ProductSummaryCustom({
         query,
         onClick: autocompleteSummary ? actionOnClick : undefined,
         onClickCapture: autocompleteSummary ? undefined : actionOnClick,
-        'aria-label': `View product details for ${
-          product?.productName || 'product'
-        }`,
+        'aria-label': `${product?.productName || 'product'}`,
       }
 
   const eventParameters =