+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
  ## [Unreleased]
 
+ ### Fixed
+
+ - Change aria-label in shelf to be the same as product name (visible label)
+ 
 ## [2.91.1] - 2026-01-07
 
 ### Added